+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 ### Bugfixes
 
 - [#4381](https://github.com/influxdata/telegraf/issues/4381): Use localhost as default server tag in zookeeper input.
+- [#4374](https://github.com/influxdata/telegraf/issues/4374): Don't set values when pattern doesn't match in regex processor.
 
 ## v1.7.1 [2018-07-03]